backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shraddhabhandarkar3 <bhandarkar.s@northeastern.edu>"]
readme = "README.md"
package-mode=false 

[tool.poetry.dependencies]
python = "3.10.15"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain-community = "^0.3.10"
serpapi = "^0.1.5"
pinecone-client = "^5.0.1"
semantic-router = "^0.0.72"
openai = "^1.57.2"
google-search-results = "^2.4.2"
boto3 = "^1.35.77"
langchain-core = "^0.3.22"
python-multipart = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
